Moved and adjusted tests, hopefully this structure is easier to understand and gives a better overview
