+++ tests/protocols/entities/tasks/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2018-2021 Greenbone Networks GmbH
+# Copyright (C) 2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -15,19 +15,3 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-from gvm.errors import GvmError
-
-
-class GmpStartTaskTestCase:
-    def test_start_task(self):
-        self.gmp.start_task('a1')
-
-        self.connection.send.has_been_called_with('<start_task task_id="a1"/>')
-
-    def test_missing_id(self):
-        with self.assertRaises(GvmError):
-            self.gmp.start_task(None)
-
-        with self.assertRaises(GvmError):
-            self.gmp.start_task('')
