+++ tests/protocols/entities/tasks/test_create_container_task.py
@@ -18,8 +18,11 @@
 
 from gvm.errors import RequiredArgument
 
+from ...gmpv208 import Gmpv208TestCase
+from ...gmpv214 import Gmpv214TestCase
 
-class GMPCreateContainerTaskCommandTestCase:
+
+class GmpCreateContainerTaskTestCase:
     def test_create_task(self):
         self.gmp.create_container_task(name='foo')
 
@@ -47,3 +50,16 @@ class GMPCreateContainerTaskCommandTestCase:
             '<comment>bar</comment>'
             '</create_task>'
         )
+
+
+# For new versions add another Mixin here.
+class Gmpv208CreateContainerTaskTestCase(
+    GmpCreateContainerTaskTestCase, Gmpv208TestCase
+):
+    pass
+
+
+class Gmpv214CreateContainerTaskTestCase(
+    GmpCreateContainerTaskTestCase, Gmpv214TestCase
+):
+    pass
